refactor(mesh): give announcement building and merging their own module
Closes #1674. `gossip.rs` was at 1,984 lines against the repository's 2,000
threshold, so the next change to it had to split something first. It also held
two responsibilities: the gossip transport, and everything about announcements.

Move the announcement half to `mesh/announcements.rs`: building a node's own
`PeerAnnouncement` and the rebroadcast set, merging direct and transitive
announcements back into `PeerInfo`, and deciding when a peer changed
meaningfully. `gossip.rs` keeps connections, streams, the join probe and the
peer lifecycle, and drops to 1,440 lines.

No behavior changes and no wire-protocol changes. The moved code is unchanged:
diffing each block against its origin leaves nothing. Only visibility moved
with it, and only where the compiler required it: two fields of
`RebroadcastAnnouncements` are now `pub(super)` because `gossip.rs` destructures
it. `LocalAnnouncementData` keeps its fields private, since both its producer
and its consumer moved together.

The tests stay in `tests/gossip/merge_and_refresh.rs` and still cover this code.
Moving them would mean extracting `test_addr`, `test_endpoint_id` and
`test_announcement` out of `tests/gossip.rs`, which eight other test files reach
through `include!`; that is a separate job from this move, and doing it here
would make the diff harder to verify rather than easier.
